requirements.txt
flask==3.0.3
flask-login==0.6.3
werkzeug==3.0.3
psycopg2-binary==2.9.9
openpyxl==3.1.2
pandas==2.2.2
gunicorn==22.0.0
